refactor(hindsight): isolate per-venue decode strategies
One netting pass tried to serve every venue, with Relay handling
(the client-fee correction, solver rebalancing fills) mixed into the
generic path. Matching now picks an explicit strategy per transaction
— sender netting, maker-finding, or Relay — and each venue's logic
lives in its own module under decoder/venues/. Adding a venue means
adding a module, not threading more special cases through the shared
loop.

Behavior change: the client-fee correction now applies only to
Relay-entered transactions. It previously ran for every venue, though
only Relay's collector was registered, so other venues' results are
unaffected in practice.
