+++ tools/hindsight/src/decoder/net.rs
@@ -64,40 +64,6 @@ pub(crate) fn decode_trade(
     net_trade(&sent, &received)
 }
 
-/// Total of each token transferred to a known client fee-collector within the transaction, keyed
-/// by token (native ETH is [`Address::ZERO`]).
-///
-/// A client like Relay skims its fee by sending part of the input token to a fee collector before
-/// swapping, so the user's netted `amount_in` includes money that never entered the swap. Backing
-/// that fee out lets the re-solve compare Fynd against the client on the amount actually routed,
-/// rather than crediting Fynd with the client's fee. Matches by recipient regardless of sender, so
-/// it catches both a direct user skim and a router skim.
-pub(crate) fn fee_to_collectors(
-    logs: &[Log],
-    native_transfers: &[(Address, Address, U256)],
-    fee_collectors: &HashSet<Address>,
-) -> HashMap<Address, U256> {
-    let mut fees: HashMap<Address, U256> = HashMap::new();
-    if fee_collectors.is_empty() {
-        return fees;
-    }
-    for &(_, to, value) in native_transfers {
-        if fee_collectors.contains(&to) {
-            *fees.entry(Address::ZERO).or_default() += value;
-        }
-    }
-    for log in logs {
-        let primitive = to_primitive_log(log);
-        let Ok(transfer) = Transfer::decode_log(&primitive) else {
-            continue;
-        };
-        if fee_collectors.contains(&transfer.to) {
-            *fees.entry(log.address()).or_default() += transfer.value;
-        }
-    }
-    fees
-}
-
 /// Net the sent and received balances into a single swap.
 ///
 /// Returns `None` unless exactly one token nets out and exactly one token nets in. A net with more
@@ -138,124 +104,6 @@ fn net_trade(sent: &HashMap<Address, U256>, received: &HashMap<Address, U256>) -
         .iter()
         .next()
         .filter(|_| net_received.len() == 1)?;
-    Some(NetSwap { token_in, amount_in, token_out, amount_out })
-}
-
-/// Decode a Relay solver-initiated rebalancing fill, where `tx.from` is a rotating solver EOA with
-/// no net flow (so [`decode_trade`] finds nothing) and the swap moves Relay's own liquidity.
-///
-/// Anchors on the fee collector, which always funds the input: `token_in` is the single token it
-/// net-sends. The output is either the token that returns to the collector (an **internal**
-/// inventory rebalance) or the asset received by the single external **pure-sink** recipient — an
-/// address that receives but never sends — for a cross-chain order fill.
-///
-/// Returns `None` (declines) when the shape is ambiguous: not exactly one input token, a same-token
-/// "swap", more than one token back to the collector, or more than one external recipient/output
-/// (a batched multi-order fill, like `net_trade`'s multi-leg guard).
-pub(crate) fn decode_relay_rebalance(
-    logs: &[Log],
-    native_transfers: &[(Address, Address, U256)],
-    fee_collectors: &HashSet<Address>,
-    relay_routers: &HashSet<Address>,
-    wrapped_native: Address,
-) -> Option<NetSwap> {
-    let mut sent: HashMap<(Address, Address), U256> = HashMap::new();
-    let mut received: HashMap<(Address, Address), U256> = HashMap::new();
-    let mut senders: HashSet<Address> = HashSet::new();
-
-    for &(from, to, value) in native_transfers {
-        senders.insert(from);
-        *sent
-            .entry((from, Address::ZERO))
-            .or_default() += value;
-        *received
-            .entry((to, Address::ZERO))
-            .or_default() += value;
-    }
-    for log in logs {
-        let Ok(transfer) = Transfer::decode_log(&to_primitive_log(log)) else {
-            continue;
-        };
-        let token = log.address();
-        senders.insert(transfer.from);
-        *sent
-            .entry((transfer.from, token))
-            .or_default() += transfer.value;
-        *received
-            .entry((transfer.to, token))
-            .or_default() += transfer.value;
-    }
-
-    // Aggregate the fee collector(s)' flow per token.
-    let mut fc_sent: HashMap<Address, U256> = HashMap::new();
-    let mut fc_recv: HashMap<Address, U256> = HashMap::new();
-    for (&(addr, token), &v) in &sent {
-        if fee_collectors.contains(&addr) {
-            *fc_sent.entry(token).or_default() += v;
-        }
-    }
-    for (&(addr, token), &v) in &received {
-        if fee_collectors.contains(&addr) {
-            *fc_recv.entry(token).or_default() += v;
-        }
-    }
-
-    // token_in: the single token the collector net-sends.
-    let net_in: Vec<(Address, U256)> = fc_sent
-        .iter()
-        .filter_map(|(&token, &s)| {
-            let r = fc_recv
-                .get(&token)
-                .copied()
-                .unwrap_or_default();
-            (s > r).then_some((token, s - r))
-        })
-        .collect();
-    if net_in.len() != 1 {
-        return None;
-    }
-    let (token_in, amount_in) = net_in[0];
-
-    // C2 internal rebalance: the collector net-receives exactly one (different) token.
-    let net_recv: Vec<(Address, U256)> = fc_recv
-        .iter()
-        .filter_map(|(&token, &r)| {
-            let s = fc_sent
-                .get(&token)
-                .copied()
-                .unwrap_or_default();
-            (r > s).then_some((token, r - s))
-        })
-        .collect();
-    if !net_recv.is_empty() {
-        if net_recv.len() != 1 || net_recv[0].0 == token_in {
-            return None;
-        }
-        let (token_out, amount_out) = net_recv[0];
-        return Some(NetSwap { token_in, amount_in, token_out, amount_out });
-    }
-
-    // C1 external fill: the single pure-sink recipient (receives but never sends), excluding
-    // infrastructure (routers, collector, the wrapped-native token, the zero address) and the
-    // input token.
-    let mut outputs: Vec<(Address, Address, U256)> = Vec::new(); // (recipient, token_out, amount)
-    for (&(addr, token), &v) in &received {
-        if token == token_in || v.is_zero() || senders.contains(&addr) {
-            continue;
-        }
-        if relay_routers.contains(&addr) ||
-            fee_collectors.contains(&addr) ||
-            addr == Address::ZERO ||
-            addr == wrapped_native
-        {
-            continue;
-        }
-        outputs.push((addr, token, v));
-    }
-    if outputs.len() != 1 {
-        return None;
-    }
-    let (_, token_out, amount_out) = outputs[0];
     Some(NetSwap { token_in, amount_in, token_out, amount_out })
 }
 
@@ -327,126 +175,6 @@ mod tests {
 
         let result = decode_trade(&logs, &native, user).unwrap();
         assert_eq!(result, swap(Address::ZERO, 1000, token, 2000));
-    }
-
-    #[test]
-    fn fee_to_collectors_totals_input_skim() {
-        let user = addr(1);
-        let router = addr(2);
-        let collector = addr(99);
-        let token_in = addr(10);
-        let pool = addr(50);
-        let collectors = HashSet::from([collector]);
-
-        // Router skims part of the input token to the collector; the rest goes to the pool.
-        let logs = vec![
-            make_transfer_log(token_in, user, router, U256::from(1000)),
-            make_transfer_log(token_in, router, collector, U256::from(40)),
-            make_transfer_log(token_in, router, pool, U256::from(960)),
-        ];
-        let fees = fee_to_collectors(&logs, &[], &collectors);
-        assert_eq!(fees.get(&token_in).copied(), Some(U256::from(40)));
-    }
-
-    #[test]
-    fn fee_to_collectors_totals_output_skim() {
-        let user = addr(1);
-        let pool = addr(50);
-        let collector = addr(99);
-        let token_out = addr(11);
-        let collectors = HashSet::from([collector]);
-
-        // Pool sends the output; part is skimmed to the collector, the rest to the user. The fee
-        // map keys this by token_out so the decoder can add it back to the gross swap output.
-        let logs = vec![
-            make_transfer_log(token_out, pool, collector, U256::from(30)),
-            make_transfer_log(token_out, pool, user, U256::from(970)),
-        ];
-        let fees = fee_to_collectors(&logs, &[], &collectors);
-        assert_eq!(fees.get(&token_out).copied(), Some(U256::from(30)));
-    }
-
-    #[test]
-    fn fee_to_collectors_empty_set_is_noop() {
-        let logs = vec![make_transfer_log(addr(10), addr(1), addr(99), U256::from(40))];
-        assert!(fee_to_collectors(&logs, &[], &HashSet::new()).is_empty());
-    }
-
-    #[test]
-    fn relay_rebalance_external_token_fill() {
-        // Collector funds token_in -> pool -> token_out delivered to an external recipient.
-        let fee = addr(99);
-        let pool = addr(50);
-        let recipient = addr(7);
-        let token_in = addr(10);
-        let token_out = addr(11);
-        let collectors = HashSet::from([fee]);
-        let routers = HashSet::from([addr(2)]);
-        let logs = vec![
-            make_transfer_log(token_in, fee, pool, U256::from(1000)),
-            make_transfer_log(token_out, pool, recipient, U256::from(2000)),
-        ];
-        let got = decode_relay_rebalance(&logs, &[], &collectors, &routers, addr(200)).unwrap();
-        assert_eq!(got, swap(token_in, 1000, token_out, 2000));
-    }
-
-    #[test]
-    fn relay_rebalance_external_native_eth_out() {
-        // C1 with native-ETH output: collector sends token_in, router delivers ETH to recipient.
-        let fee = addr(99);
-        let router = addr(2);
-        let pool = addr(50);
-        let recipient = addr(7);
-        let token_in = addr(10);
-        let collectors = HashSet::from([fee]);
-        let routers = HashSet::from([router]);
-        let logs = vec![make_transfer_log(token_in, fee, pool, U256::from(1000))];
-        let native = vec![(router, recipient, U256::from(2000))];
-        let got = decode_relay_rebalance(&logs, &native, &collectors, &routers, addr(200)).unwrap();
-        assert_eq!(got, swap(token_in, 1000, Address::ZERO, 2000));
-    }
-
-    #[test]
-    fn relay_rebalance_internal_back_to_collector() {
-        // C2: collector sends token_in, receives token_out back (inventory rebalance).
-        let fee = addr(99);
-        let pool = addr(50);
-        let token_in = addr(10);
-        let token_out = addr(11);
-        let collectors = HashSet::from([fee]);
-        let routers = HashSet::from([addr(2)]);
-        let logs = vec![
-            make_transfer_log(token_in, fee, pool, U256::from(1000)),
-            make_transfer_log(token_out, pool, fee, U256::from(1001)),
-        ];
-        let got = decode_relay_rebalance(&logs, &[], &collectors, &routers, addr(200)).unwrap();
-        assert_eq!(got, swap(token_in, 1000, token_out, 1001));
-    }
-
-    #[test]
-    fn relay_rebalance_declines_multi_recipient() {
-        // One input token but two external recipients = batched multi-order fill -> decline.
-        let fee = addr(99);
-        let pool = addr(50);
-        let token_in = addr(10);
-        let token_out = addr(11);
-        let collectors = HashSet::from([fee]);
-        let routers = HashSet::from([addr(2)]);
-        let logs = vec![
-            make_transfer_log(token_in, fee, pool, U256::from(2000)),
-            make_transfer_log(token_out, pool, addr(7), U256::from(1000)),
-            make_transfer_log(token_out, pool, addr(8), U256::from(1000)),
-        ];
-        assert!(decode_relay_rebalance(&logs, &[], &collectors, &routers, addr(200)).is_none());
-    }
-
-    #[test]
-    fn relay_rebalance_declines_without_collector_outflow() {
-        // No transfer from a fee collector -> nothing to anchor on.
-        let logs = vec![make_transfer_log(addr(10), addr(1), addr(50), U256::from(1000))];
-        let collectors = HashSet::from([addr(99)]);
-        let routers = HashSet::from([addr(2)]);
-        assert!(decode_relay_rebalance(&logs, &[], &collectors, &routers, addr(200)).is_none());
     }
 
     #[test]
